Add complete userspace tooling and enhanced shell (#83)
* Added change directory command

* Added /bin/cd to the root mount

* Ignored cd

* Created cd & pwd as builtin

* Moved fcntl to uapi

* Created the touch binary

* Ignoring touch binary

* Added the cat command

* Removed print

* Added rm, rmdir & mkdir

* Added unix env

* Added unlink to userspc & updated Makefile

* Added a basic echo

* Added time syscall to userspace

* Updated time.c

* Implemented reboot in userspace

* Minor tweak

* Added modules and mount

* Added insmod, rmmod and mount

* Added a simple malloc. NOSYS though

* Added BRK in userspace

* Removed calloc & realloc

* Ignore certain binaries

+++ kernel/include/uapi/types.h
@@ -14,6 +14,6 @@ typedef unsigned short uid_t;
 typedef unsigned short gid_t;
 
 typedef unsigned int size_t;
-typedef unsigned long long time_t;
+typedef signed long time_t;
 
 #endif
